lambda/thumbnail/requirements.txt
Pillow
pdf2image
